refactor: Auth handlers and add service layer
