--- marketplace-be/handlers/auth.go
@@ -1,77 +1,0 @@
-package handlers
-
-import (
-	"log"
-	"net/http"
-
-	"marketplace-be/auth"
-	"marketplace-be/database"
-	"marketplace-be/models"
-
-	"github.com/gin-gonic/gin"
-	"golang.org/x/crypto/bcrypt"
-)
-
-func Signup(c *gin.Context) {
-	var input models.SignupInput
-	if err := c.ShouldBindJSON(&input); err != nil {
-		log.Printf("Error binding JSON: %v", err)
-		c.JSON(http.StatusBadRequest, gin.H{"message": "Invalid input format"})
-		return
-	}
-
-	var user models.User
-	if err := database.DB.Where("email = ?", input.Email).First(&user).Error; err == nil {
-		log.Printf("Attempted signup with already registered email: %s", input.Email)
-		c.JSON(http.StatusConflict, gin.H{"message": "Email already registered"})
-		return
-	}
-
-	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(input.Password), bcrypt.DefaultCost)
-	if err != nil {
-		log.Printf("Error hashing password: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"message": "Internal Server Error"})
-		return
-	}
-
-	newUser := models.User{
-		Name:         input.Name,
-		Email:        input.Email,
-		PasswordHash: string(hashedPassword),
-	}
-
-	if err := database.DB.Create(&newUser).Error; err != nil {
-		log.Printf("Error creating user: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"message": "Internal Server Error"})
-		return
-	}
-
-	c.JSON(http.StatusCreated, gin.H{"message": "User created successfully"})
-}
-
-func Login(c *gin.Context) {
-	var input models.LoginInput
-	if err := c.ShouldBindJSON(&input); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"message": "Invalid input format"})
-		return
-	}
-
-	var user models.User
-	if err := database.DB.Where("email = ?", input.Email).First(&user).Error; err != nil {
-		c.JSON(http.StatusNotFound, gin.H{"message": "Invalid credentials, try again"})
-		return
-	}
-
-	if err := bcrypt.CompareHashAndPassword([]byte(user.PasswordHash), []byte(input.Password)); err != nil {
-		c.JSON(http.StatusUnauthorized, gin.H{"message": "Invalid credentials, try again"})
-		return
-	}
-
-	token, err := auth.GenerateToken(user.ID)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"message": "Failed to generate token"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"token": token})
-}
